BLRBT-8: (WIP) Fix Gulp `watch` task
This commit allows the Gulp `watch` task to work from within a Vagrant
environment, with the exception of the `build/css` task. I haven't yet
determined what, specifically, is causing the issue but when one of the
source `.scss` files is updated the first time, it triggers the CSS
build task normally. The 2nd time, the CSS build task is triggered
multiple times, along with all of the other build tasks, and this
continues until the process is killed either manually or automatically.